refactor: replace recent domains section with a dedicated component in SignUp page
