+++ app/client/src/pages/UserAuth/SignUp.tsx
@@ -66,7 +66,7 @@ import { isLoginHostname } from "utils/cloudBillingUtils";
 import { appsmithTelemetry } from "instrumentation";
 import { getIsAiAgentInstanceEnabled } from "ee/selectors/aiAgentSelectors";
 import { getSafeErrorMessage } from "ee/constants/approvedErrorMessages";
-import { getRecentDomains, isValidAppsmithDomain } from "utils/multiOrgDomains";
+import RecentDomainsSection from "./RecentDomainsSection";
 
 declare global {
   interface Window {
@@ -76,7 +76,6 @@ declare global {
   }
 }
 const { cloudHosting, googleRecaptchaSiteKey } = getAppsmithConfigs();
-const recentDomains = getRecentDomains();
 
 const validate = (values: SignupFormValues) => {
   const errors: SignupFormValues = {};
@@ -95,63 +94,6 @@ const validate = (values: SignupFormValues) => {
 
   return errors;
 };
-
-const recentDomainsSection = recentDomains.length > 0 && (
-  <div className="mt-8">
-    <div className="mb-2">
-      <Text kind="body-m">{createMessage(YOU_VE_ALREADY_SIGNED_INTO)}</Text>
-    </div>
-
-    <div className="max-h-48 overflow-y-auto border border-gray-200 rounded-md py-4 px-3">
-      {recentDomains.map((domain, index) => {
-        const orgName = domain
-          .split(".")[0]
-          .split("-")
-          .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-          .join(" ");
-
-        const avatarLetter = String.fromCharCode(65 + (index % 26));
-
-        const isLastItem = index === recentDomains.length - 1;
-
-        return (
-          <div
-            className={`flex items-center justify-between p-1 ${
-              isLastItem ? "mb-0" : "mb-3"
-            }`}
-            key={domain}
-          >
-            <div className="flex items-center space-x-3">
-              <div className="w-8 h-8 bg-[color:var(--ads-color-background-secondary)] rounded-full flex items-center justify-center text-gray-600 font-light text-sm">
-                {avatarLetter}
-              </div>
-              <div className="flex flex-col">
-                <span className="text-md font-semibold text-gray-700 max-w-[180px] line-clamp-1">
-                  {orgName}
-                </span>
-                <span className="text-xs font-light text-gray-500 max-w-[180px] line-clamp-1">
-                  {domain}
-                </span>
-              </div>
-            </div>
-            <Button
-              className="px-4 py-2 text-sm"
-              kind="secondary"
-              onClick={() => {
-                if (isValidAppsmithDomain(domain)) {
-                  window.location.href = `https://${domain}/user/login`;
-                }
-              }}
-              size="md"
-            >
-              Open
-            </Button>
-          </div>
-        );
-      })}
-    </div>
-  </div>
-);
 
 type SignUpFormProps = InjectedFormProps<
   SignupFormValues,
@@ -372,7 +314,9 @@ export function SignUp(props: SignUpFormProps) {
         </SpacedSubmitForm>
       )}
       {isCloudBillingEnabled && isHostnameEqualtoLogin && cloudBillingSignIn}
-      {isCloudBillingEnabled && isHostnameEqualtoLogin && recentDomainsSection}
+      {isCloudBillingEnabled && isHostnameEqualtoLogin && (
+        <RecentDomainsSection />
+      )}
     </Container>
   );
 }
